Guard against mutating the passed object
We need to clone the object internally, otherwise we create side effects outside our function.

I expect only `realChangesetParser`  to be used outside, so cloning here is important. However since `realChangesetElementParser` is also helpful, there are now two versions, one that does it's own cloning and one that doesn't and expect the cloning to be done outside (like with `realChangesetParser`). Only one is exported via the index file to keep the public API small.

At least I hope that is how it works :-D